+++ uk-pmrv-app-api/src/main/java/uk/gov/pmrv/api/notification/template/repository/impl/DocumentTemplateCustomRepositoryImpl.java
@@ -72,7 +72,7 @@ public class DocumentTemplateCustomRepositoryImpl implements DocumentTemplateCus
 
         if(!forCount) {
             query.setParameter("limit", searchCriteria.getPaging().getPageSize());
-            query.setParameter("offset", searchCriteria.getPaging().getPageNumber() * searchCriteria.getPaging().getPageSize());
+            query.setParameter("offset", (long) searchCriteria.getPaging().getPageNumber() * searchCriteria.getPaging().getPageSize());
         }
 
         return query;
