+++ uk-pmrv-app-api/src/main/java/uk/gov/pmrv/api/notification/template/repository/impl/NotificationTemplateCustomRepositoryImpl.java
@@ -75,7 +75,7 @@ public class NotificationTemplateCustomRepositoryImpl implements NotificationTem
 
         if(!forCount) {
             query.setParameter("limit", searchCriteria.getPaging().getPageSize());
-            query.setParameter("offset", searchCriteria.getPaging().getPageNumber() * searchCriteria.getPaging().getPageSize());
+            query.setParameter("offset", (long) searchCriteria.getPaging().getPageNumber() * searchCriteria.getPaging().getPageSize());
         }
 
         return query;
