+++ uk-pmrv-app-api/src/main/java/uk/gov/pmrv/api/workflow/request/application/item/repository/ItemRegulatorRepository.java
@@ -49,7 +49,7 @@ public class ItemRegulatorRepository {
                 .where(constructWherePredicate(userId, assignmentType, requestTask, request,
                 		scopedCARequestTaskTypes))
                 .orderBy(requestTask.startDate.desc())
-                .offset(paging.getPageNumber() * paging.getPageSize())
+                .offset((long) paging.getPageNumber() * paging.getPageSize())
                 .limit(paging.getPageSize());
 
         return ItemPage.builder()
